lifting state & add nav bar by functional components
